Update StateMachine Diagram
