--- src/main/packages/common/uml-state/uml-state-variable.ts
@@ -1,3 +1,0 @@
-import { UMLStateMember } from './uml-state-member';
-
-export abstract class UMLStateVariable extends UMLStateMember {}
