+++ src/main/packages/components.ts
@@ -63,7 +63,7 @@ import { BPMNGroupComponent } from './bpmn/bpmn-group/bpmn-group-component';
 import { BPMNPoolComponent } from './bpmn/bpmn-pool/bpmn-pool-component';
 import { BPMNSwimlaneComponent } from './bpmn/bpmn-swimlane/bpmn-swimlane-component';
 import { BPMNDataStoreComponent } from './bpmn/bpmn-data-store/bpmn-data-store-component';
-import { UMLStateComponent } from './common/uml-state/uml-state-component';
+import { UMLStateComponent } from './uml-state-diagram/uml-state/uml-state-component';
 import { UMLStateActionNodeComponent } from './uml-state-diagram/uml-state-action-node/uml-state-action-node-component';
 import { UMLStateFinalNodeComponent } from './uml-state-diagram/uml-state-final-node/uml-state-final-node-component';
 import { UMLStateForkNodeComponent } from './uml-state-diagram/uml-state-fork-node/uml-state-fork-node-component';
@@ -73,7 +73,7 @@ import { UMLStateMergeNodeComponent } from './uml-state-diagram/uml-state-merge-
 import { UMLStateObjectNodeComponent } from './uml-state-diagram/uml-state-object-node/uml-state-object-node-component';
 import { UMLStateTransitionComponent } from './uml-state-diagram/uml-state-transition/uml-state-transition-component';
 import { UMLStateCodeBlockComponent } from './uml-state-diagram/uml-state-code-block/uml-state-code-block-component';
-import { UMLStateMemberComponent } from './common/uml-state/uml-state-member-component';
+import { UMLStateMemberComponent } from './uml-state-diagram/uml-state/uml-state-member-component';
 
 export const Components: {
   [key in UMLElementType | UMLRelationshipType]:
@@ -160,9 +160,9 @@ export const Components: {
   [UMLRelationshipType.SyntaxTreeLink]: SyntaxTreeLinkComponent,
   [UMLRelationshipType.FlowchartFlowline]: FlowchartFlowlineComponent,
   [UMLRelationshipType.BPMNFlow]: BPMNFlowComponent,
-  [UMLElementType.State_]: UMLStateComponent,
-  [UMLElementType.StateVariable_]: UMLStateMemberComponent,
-  [UMLElementType.StateAction_]: UMLStateMemberComponent,
+  [UMLElementType.State]: UMLStateComponent,
+  [UMLElementType.StateVariable]: UMLStateMemberComponent,
+  [UMLElementType.StateAction]: UMLStateMemberComponent,
   [UMLElementType.StateActionNode]: UMLStateActionNodeComponent,
   [UMLElementType.StateFinalNode]: UMLStateFinalNodeComponent,
   [UMLElementType.StateForkNode]: UMLStateForkNodeComponent,
