+++ src/main/packages/uml-state-diagram/uml-state-action/uml-state-action.ts
@@ -1,7 +1,6 @@
+import { UMLStateMember } from '../uml-state/uml-state-member';
 import { StateElementType } from '..';
-import { UMLStateAction } from '../../common/uml-state/uml-state-action';
-import { UMLElementType } from '../../uml-element-type';
 
-export class UMLStateAction_ extends UMLStateAction {
-  type = StateElementType.StateAction_;
-}
+export class UMLStateAction extends UMLStateMember {
+  type = StateElementType.StateAction;
+}
