+++ src/main/packages/uml-state-diagram/uml-state-variable/uml-state-variable-update.tsx
