+++ src/main/packages/uml-state-diagram/uml-state-variable/uml-state-variable.ts
@@ -1,7 +1,6 @@
+import { UMLStateMember } from '../uml-state/uml-state-member';
 import { StateElementType } from '..';
-import { UMLStateVariable } from '../../common/uml-state/uml-state-variable';
-import { UMLElementType } from '../../uml-element-type';
 
-export class UMLStateVariable_ extends UMLStateVariable {
-  type: UMLElementType = StateElementType.StateVariable_;
-}
+export class UMLStateVariable extends UMLStateMember {
+  type = StateElementType.StateVariable;
+}
