+++ src/main/packages/uml-state-diagram/uml-state/uml-state-member.ts
